+++ sources/BranchSitePath.h
@@ -93,11 +93,11 @@ class BranchSitePath  {
 		return (i<j) ? (2 * nstate - i - 1) * i / 2 + j - i - 1 : (2 * nstate - j - 1) * j / 2 + i - j - 1 ;
 	}
 
-	void AddRateSuffStat(int& count, double& beta, double factor, const double* rr, const double* stat, int nstate);
-	void AddProfileSuffStat(int* count, double* beta, double factor, const double* rr, int nstate);
-	void AddRRSuffStat(int* count, double* beta, double factor, const double* stat, int nstate);
+	void AddRateSuffStat(double& count, double& beta, double factor, const double* rr, const double* stat, int nstate);
+	void AddProfileSuffStat(double* count, double* beta, double factor, const double* rr, int nstate);
+	void AddRRSuffStat(double* count, double* beta, double factor, const double* stat, int nstate);
 
-	void AddGeneralPathRateSuffStat(int& count, double& beta, double factor, SubMatrix* mat);
+	void AddGeneralPathRateSuffStat(double& count, double& beta, double factor, SubMatrix* mat);
 	void AddGeneralPathSuffStat(map<pair<int,int>,int>& paircount, map<int,double>& waitingtime, double factor);
 
 	Plink* init;
