+++ sources/GeneralPathSuffStatMatrixSubstitutionProcess.h
@@ -28,8 +28,8 @@ class GeneralPathSuffStatMatrixSubstitutionProcess : public virtual MatrixSubsti
 
 	protected:
 
-	void AddBranchLengthSuffStat(int& count, double& beta, BranchSitePath** patharray, int* nonmissing);
-	void AddSiteRateSuffStat(int* count, double* beta, BranchSitePath** patharray, double length, int* nonmissing);
+	void AddBranchLengthSuffStat(double& count, double& beta, BranchSitePath** patharray, int* nonmissing);
+	void AddSiteRateSuffStat(double* count, double* beta, BranchSitePath** patharray, double length, int* nonmissing);
 	void AddSiteProfileSuffStat(int* siterootstate, map<pair<int,int>, int>* sitepaircount, map<int,double>* sitewaitingtime, BranchSitePath** patharray, double branchlength, int* nonmissing);
 
 };
